requirements.txt
gymnasium[mujoco]>=0.29.0
numpy>=1.24.0
torch>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0


